moved updateProduct action and thunk creators to singleProduct reducer
