+++ client/components/EditProduct/editProduct.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {Redirect} from 'react-router-dom'
-import {updateProductOnServer} from '../../reducer/allProds'
+import {updateProductOnServer} from '../../reducer/singleProduct'
 
 class EditProduct extends Component {
   constructor() {
